Display the daily electricity dates in UTC
The dates in the Daily Electricities list are now displayed in UTC.
Before, the dates were converted to locale date. This was not correct.
In fact, the REST endpoint saves the dates in UTC. When the dates are
retrieved, and contain time, it is safe to convert them to locale date.
Otherwise, the dates should be displayed as is. If an attempt is made
to convert a date without time to locale date, wrong results will
be obtained as was the case with the Daily Electricities list.