Correct the selection bug introduced by the sorting commit
Before introducing column sorting in a last commit, during edits and
deletes, panels were selected based on their indices in the list. That
initially caused no problem. But with column sorting introduced, only
the visual appearance of the data to the user changed. The underlying
indices remained the same. This means that in sorting a column in
descending order, and selecting the first row, a user could actually be
selecting the element at index 0 which is different to the one that
appears at first position on the screen. The element appearing at the
first position on the screen is the last element because the list is
sorted in descending order, but the actual underlying element at index
0 is not the one selected.