Get hourly and daily electricities by panel
Before, the PanelService was exposing methods that required URIs to
get hourly and daily electricities. That forced the controller to know
that a URI is required to get electricities. We want the controller to
think in terms of objects. Passing a panel to a method that will get
hourly or daily electricities is more intuitive.